+++ neoforge-1.21.1/src/main/java/com/createtiers/foundation/item/TieredKineticStats.java
@@ -68,7 +68,7 @@ public class TieredKineticStats implements TooltipModifier {
                 .addTo(tooltip);
 
         if (block instanceof TieredShaftBlock) {
-            ModLang.translate("tooltip.tier_calibration_fallback")
+            ModLang.translate("tooltip.in_world_tier_upgrade")
                     .style(ChatFormatting.DARK_GRAY)
                     .addTo(tooltip);
         }